Bug fixes - Do not delete unknow IDs at given intervals - Calculate last lap for runners with too many pass times

+++ LiveResults.Client/FrmETimingMonitor.cs
@@ -73,7 +73,7 @@ namespace LiveResults.Client
                 var dbRunners = client.m_runners.Values;
                 foreach (var dbRunner in dbRunners.ToList())
                 {
-                    if (!usedIds.Contains(dbRunner.ID))
+                    if (!usedIds.Contains(dbRunner.ID) && dbRunner.ID >= 0)
                         client.DeleteID(dbRunner.ID);
                 }
             }
